@@ -51,42 +51,39 @@
   "author": "dgreif",
   "license": "MIT",
   "dependencies": {
+    "@homebridge/camera-utils": "1.2.0",
     "agentkeepalive": "4.1.3",
     "cacheable-lookup": "5.0.3",
     "colors": "^1.4.0",
     "debug": "^4.1.1",
-    "execa": "^4.0.3",
-    "ffmpeg-for-homebridge": "0.0.7",
-    "get-port": "^5.1.1",
-    "got": "^11.5.2",
+    "got": "^11.6.0",
     "node-machine-id": "^1.1.12",
-    "rxjs": "^6.6.2",
+    "rxjs": "^6.6.3",
     "sdp": "^3.0.2",
     "sip": "0.0.6",
     "socket.io": "^2.3.0",
     "stun": "2.1.0",
-    "systeminformation": "4.27.3",
     "uuid": "^8.3.0"
   },
   "devDependencies": {
-    "@nestjs/common": "7.4.3",
-    "@nestjs/core": "7.4.3",
+    "@nestjs/common": "7.4.4",
+    "@nestjs/core": "7.4.4",
     "@types/debug": "^4.1.5",
-    "@types/jest": "^26.0.10",
-    "@types/node": "^14.6.1",
+    "@types/jest": "^26.0.13",
+    "@types/node": "^14.6.4",
     "@types/socket.io-client": "^1.4.33",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^3.10.1",
     "@typescript-eslint/parser": "^3.10.1",
     "conventional-github-releaser": "^3.1.5",
     "dotenv": "^8.2.0",
-    "eslint": "^7.7.0",
+    "eslint": "^7.8.1",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-jest": "^23.20.0",
+    "eslint-plugin-jest": "^24.0.0",
     "eslint-plugin-prettier": "^3.1.4",
     "express": "^4.17.1",
     "homebridge": "1.1.2",
-    "homebridge-config-ui-x": "^4.27.0",
+    "homebridge-config-ui-x": "^4.27.1",
     "jest": "^26.4.2",
     "prettier": "^2.1.1",
     "reflect-metadata": "0.1.13",
